requirements.txt
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Cache==0.13.1
Flask-Caching==1.9.0
Flask-Cors==3.0.9
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
moviepy==1.0.3
spleeter==1.5.4
Werkzeug==1.0.1
youtube-dl==2020.9.14
youtube-transcript-api==0.3.1
lyricsgenius==2.0.0